requirements.txt
temporalio>=1.7.0
motor>=3.6.0
pymongo>=4.10.0
voyageai>=0.3.0
anthropic>=0.40.0
python-dotenv>=1.0.0
